feat(controllers): updated endpts w header valdiation

+++ src/main/java/com/greenloop/event_service/controllers/EventController.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import java.util.*;
 
+import com.greenloop.event_service.exceptions.RoleNotAllowedException;
 import com.greenloop.event_service.models.Event;
 import com.greenloop.event_service.services.EventService;
 
@@ -30,7 +31,10 @@ public class EventController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
-    public Event createEvent(@RequestBody Event event) {
+    public Event createEvent(@RequestBody Event event, @RequestHeader("X-User-Role") String userRole) {
+        if (!userRole.equals("ADMIN")) {
+            throw new RoleNotAllowedException();
+        }
         return eventService.createEvent(event);
     }
 
@@ -45,35 +49,44 @@ public class EventController {
     }
 
     @GetMapping(value = "/{id}/qr", produces = "image/png")
-    public @ResponseBody byte[] getEventQr(@PathVariable UUID id) {
+    public @ResponseBody byte[] getEventQr(@PathVariable UUID id, @RequestHeader("X-User-Role") String userRole) {
+        if (!userRole.equals("ADMIN")) {
+            throw new RoleNotAllowedException();
+        }
         // default size 300x300
         return eventService.getQrCodeImage(id, 300, 300);
     }
 
     @PutMapping("/{id}")
-    public Event updateEvent(@PathVariable UUID id, @Valid @RequestBody Event event) {
+    public Event updateEvent(@PathVariable UUID id, @Valid @RequestBody Event event, @RequestHeader("X-User-Role") String userRole) {
+        if (!userRole.equals("ADMIN")) {
+            throw new RoleNotAllowedException();
+        }
         return eventService.updateEvent(id, event);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteEvent(@PathVariable UUID id) {
+    public void deleteEvent(@PathVariable UUID id, @RequestHeader("X-User-Role") String userRole) {
+        if (!userRole.equals("ADMIN")) {
+            throw new RoleNotAllowedException();
+        }
         eventService.deleteEvent(id);
     }
 
     // get upcoming events for user
     @GetMapping("/upcoming/joined")
-    public List<Event> upcomingEventsForUser(@RequestHeader UUID userId) {
-        return eventService.upcomingEventForUser(userId);
+    public List<Event> upcomingEventsForUser(@RequestHeader("X-User-ID") String userId) {
+        return eventService.upcomingEventForUser(UUID.fromString(userId));
     }
     // get upcoming events for unjoined user
     @GetMapping("/upcoming/unjoined")
-    public List<Event> upcomingNotJoinedEvents(@RequestHeader UUID userId) {
-        return eventService.upcomingNotJoinedEvents(userId);
+    public List<Event> upcomingNotJoinedEvents(@RequestHeader("X-User-ID") String userId) {
+        return eventService.upcomingNotJoinedEvents(UUID.fromString(userId));
     }
     // get past events for user
     @GetMapping("/past")
-    public List<Event> pastEventsForUser(@PathVariable UUID eventId, @RequestHeader UUID userId) {
-        return eventService.pastEventsForUser(userId);
+    public List<Event> pastEventsForUser(@PathVariable UUID eventId, @RequestHeader("X-User-ID") String userId) {
+        return eventService.pastEventsForUser(UUID.fromString(userId));
     }
 
 }
